all warnings and memory leaks fixed

+++ code/array.cpp
@@ -12,7 +12,7 @@ void print(Line* text, size_t size, FILE* file)
 
     for(size_t y = 0; y < size; y++)
     {
-        fprintf(file, "[%lu] [%lu] ", y, text[y].lenght);
+        fprintf(file, "[%lu] [%d] ", y, text[y].lenght);
         int i = 0;
         char c;
         short unspace = 0;
@@ -82,11 +82,11 @@ size_t initialize_text(Line** text, char* buffer, size_t size)
         if(buffer[i] == '\n')
         {       
             (*text)[j].str = buffer+i+1;
-            (*text)[j-1].lenght = (size_t)((*text)[j].str - (*text)[j-1].str)-1;
+            (*text)[j-1].lenght = (int)((*text)[j].str - (*text)[j-1].str)-1;
             j++;
         }
     }
-    (*text)[count-1].lenght = size - (size_t)((*text)[count-1].str - (*text)[0].str);
+    (*text)[count-1].lenght = (int)size - (int)((*text)[count-1].str - (*text)[0].str);
 
     return count;
 }
